# speed up pmap axis-size getting