For tree_flatten, we also tried these implementations:

```python
def tree_flatten(tree):
  leaves = []
  def f(tree):
    node_type = _get_node_type(tree)
    if node_type:
      children, node_spec = node_type.to_iterable(tree)
      child_specs = tuple(map(f, children))
      return PyTreeDef(node_type, node_spec, child_specs)
    else:
      leaves.append(tree)
      return leaf
  return leaves, f(tree)

def tree_flatten(tree):
  leaves = []
  _, treedef = walk_pytree(lambda node: None, leaves.append, tree)
  return leaves, treedef
```

and surprisingly the version in this PR was slightly faster. We only tried on flattening this pytree though:

```python
pytree = [[(np.zeros(10), (), np.zeros(10), ()) for _ in range(20)] for _ in range(20)]
```

The change in this PR made `api._pmap_axis_size(pytree)` on the above `pytree` go from 5.3ms to 3.24us (notice ms and us!), and `tree_util.tree_flatten(pytree)` go from 3.5ms to 2.5ms.